debug

+++ main.h
@@ -20,7 +20,7 @@ typedef struct stpatients {
 
 typedef struct stlifestyle {
     int id[10000];
-    int physical_activity[10000];
+    float physical_activity[10000];
     float alcohol_consumption[10000];
     float caffeine_consumption[10000];
     float sleep_quality[10000];
